refactored fetch_entry
